Minor bug in test file; making sure of return functions on routes

+++ API/controllers.go
@@ -18,12 +18,14 @@ func CreateFair(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		msgErr := "ERROR: could not decode JSON file in: http://localhost:8000/api/fairs on POST request"
 		jsonErrMsg(w, msgErr)
+		return
 	}
 
 	err = models.ValidateFair(&newFair)
 	if err != nil {
 		msgErr := "ERROR: JSON file error, all fields must have a value"
 		jsonErrMsg(w, msgErr)
+		return
 	}
 
 	createFair := models.Fair{Longitude: newFair.Longitude, Latitude: newFair.Latitude, SetCen: newFair.SetCen, AreaP: newFair.AreaP, CodDist: newFair.CodDist, District: newFair.District, CodSubPref: newFair.CodSubPref, SubPref: newFair.SubPref, RegionFive: newFair.RegionFive, RegionEight: newFair.RegionEight, NameFair: newFair.NameFair, Record: newFair.Record, Street: newFair.Street, Number: newFair.Number, Neighbourhood: newFair.Neighbourhood, Reference: newFair.Reference}
@@ -33,10 +35,10 @@ func CreateFair(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		msgErr := "ERROR: could not encode JSON file in: http://localhost:8000/api/fairs on POST request"
 		jsonErrMsg(w, msgErr)
+		return
 	}
 
 	log.Println("INFO: post request http://localhost:8000/api/fairs new fair created")
-	w.WriteHeader(http.StatusCreated)
 }
 
 //DeleteFair  delete models.Fair by given id
@@ -44,18 +46,27 @@ func DeleteFair(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 	id := vars["searchParam"]
 
+	var totalFairs int64
 	var fair models.Fair
+	database.DB.Select("*").Where("id = ?", id).Table("fairs").Find(&fair).Count(&totalFairs)
+	if totalFairs < 1 {
+		msgErr := "ERROR: impossible to delete a fair that does not exist!"
+		jsonErrMsg(w, msgErr)
+		return
+	}
+
 	database.DB.Delete(&fair, id)
+
+	msg := "INFO: fair with id:" + id + " DELETED"
+	log.Println(msg)
+	_ = json.NewEncoder(w).Encode(msg)
 
 	err := json.NewEncoder(w).Encode(fair)
 	if err != nil {
 		msgErr := "ERROR: could not encode JSON file in: http://localhost:8000/api/fairs on DELETE request"
 		jsonErrMsg(w, msgErr)
+		return
 	}
-
-	msg := "INFO: fair with id:" + id + " DELETED"
-	log.Println(msg)
-	_ = json.NewEncoder(w).Encode(msg)
 }
 
 //UpdateFair update models.Fair by given id
@@ -63,19 +74,27 @@ func UpdateFair(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 	id := vars["searchParam"]
 
+	var totalFairs int64
 	var fair models.Fair
-	database.DB.Select("*").Where("id = ?", id).Table("fairs").Find(&fair)
+	database.DB.Select("*").Where("id = ?", id).Table("fairs").Find(&fair).Count(&totalFairs)
+	if totalFairs < 1 {
+		msgErr := "ERROR: impossible to update a fair that does not exist!"
+		jsonErrMsg(w, msgErr)
+		return
+	}
 
 	err := json.NewDecoder(r.Body).Decode(&fair)
 	if err != nil {
 		msgErr := "ERROR: could not decode JSON file in: http://localhost:8000/api/fairs on PUT request"
 		jsonErrMsg(w, msgErr)
+		return
 	}
 
 	err = models.ValidateFair(&fair)
 	if err != nil {
 		msgErr := "ERROR: JSON file error, all fields must have a value"
 		jsonErrMsg(w, msgErr)
+		return
 	}
 
 	//Probably a way to do it simpler
@@ -83,6 +102,7 @@ func UpdateFair(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		msgErr := "ERROR: could not convert string to int in: http://localhost:8000/api/fairs on PUT request"
 		jsonErrMsg(w, msgErr)
+		return
 	}
 
 	updateFair := models.Fair{Id: sameID, Longitude: fair.Longitude, Latitude: fair.Latitude, SetCen: fair.SetCen, AreaP: fair.AreaP, CodDist: fair.CodDist, District: fair.District, CodSubPref: fair.CodSubPref, SubPref: fair.SubPref, RegionFive: fair.RegionFive, RegionEight: fair.RegionEight, NameFair: fair.NameFair, Record: fair.Record, Street: fair.Street, Number: fair.Number, Neighbourhood: fair.Neighbourhood, Reference: fair.Reference}
@@ -92,10 +112,10 @@ func UpdateFair(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		msgErr := "ERROR: could not encode JSON file in: http://localhost:8000/api/fairs on PUT request"
 		jsonErrMsg(w, msgErr)
+		return
 	}
 
 	log.Println("INFO: put request http://localhost:8000/api/fairs fair updated with ID:" + id)
-	w.WriteHeader(http.StatusAccepted)
 }
 
 //GetFair read and find models.Fair using SearchFair
@@ -131,8 +151,9 @@ func GetFair(w http.ResponseWriter, r *http.Request) {
 			return
 		}
 	default:
-		msgErr := "WARNING: searchType: " + searchType + " unexpected. searchType must be [district,region,name or neighbourhood]"
+		msgErr := "WARNING: searchType: " + searchType + " unexpected. searchType must be [district,region,name,neighbourhood or id]"
 		jsonErrMsg(w, msgErr)
+		return
 	}
 }
 
@@ -155,7 +176,7 @@ func SearchFair(w http.ResponseWriter, searchParam, searchType, query string) er
 		return err
 	}
 
-	log.Println("INFO: get request to http://localhost:8000/api/fairs" + searchType + "/" + searchParam)
+	log.Println("INFO: get request to http://localhost:8000/api/fairs/" + searchType + "/" + searchParam)
 	return nil
 }
 
@@ -168,11 +189,13 @@ func GetFairs(w http.ResponseWriter, r *http.Request) {
 	if totalFairs < 1 {
 		msgErr := "WARNING: get request to http://localhost:8000/api/fairs	have return 0"
 		jsonErrMsg(w, msgErr)
+		return
 	} else {
 		err := json.NewEncoder(w).Encode(fairs)
 		if err != nil {
 			msgErr := "ERROR: could not encode JSON file in: http://localhost:8000/api/fairs"
 			jsonErrMsg(w, msgErr)
+			return
 		}
 
 		log.Println("INFO: get request to http://localhost:8000/api/fairs" + "	fairs found:" + strconv.Itoa(int(totalFairs)))
